requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alembic==1.13.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
asttokens==2.4.1
astunparse==1.6.3
attrs==25.1.0
bleach==6.2.0
blinker==1.7.0
bokeh==3.6.3
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.0
contourpy==1.3.1
cycler==0.11.0
datasets==3.5.0
debugpy==1.8.0
decorator==5.1.1
diffusers==0.27.0
dill==0.3.8
distro==1.9.0
dotenv==0.9.9
executing==2.0.1
fastapi==0.115.11
ffmpy==0.5.0
filelock==3.13.1
Flask==3.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
fonttools==4.42.0
frozenlist==1.5.0
fsspec==2024.3.0
gast==0.6.0
gensim==4.3.3
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.14.0
google-auth==2.23.4
google-cloud-speech==2.22.0
google-pasta==0.2.0
googleapis-common-protos==1.61.0
gradio==5.20.1
gradio_client==1.7.2
greenlet==3.0.3
groovy==0.1.2
grpcio==1.59.2
grpcio-status==1.59.2
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.1
idna==3.4
imageio==2.37.0
importlib_metadata==7.0.2
ipykernel==6.26.0
ipython==8.17.2
ipywidgets==8.1.5
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jiter==0.9.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_bokeh==4.0.5
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyterlab_widgets==3.0.13
kaggle==1.6.17
keras==3.8.0
kiwisolver==1.4.4
lazy_loader==0.4
libclang==18.1.1
linkify-it-py==2.0.3
Mako==1.3.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.2
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.3.2
multiprocess==0.70.16
namex==0.0.8
narwhals==1.29.1
nest-asyncio==1.5.8
networkx==3.1
nltk==3.9.1
numpy==1.26.4
openai==1.66.3
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.1
orjson==3.10.15
packaging==23.1
pandas==2.2.3
panel==1.6.1
param==2.2.0
parso==0.8.3
pillow==11.1.0
platformdirs==4.0.0
prompt-toolkit==3.0.41
propcache==0.3.1
proto-plus==1.22.3
protobuf==4.25.1
psutil==5.9.6
pure-eval==0.2.2
pyarrow==19.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.14
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.16.1
pyparsing==3.0.9
PySimpleGUI==4.60.5
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2024.2
pyviz_comms==3.0.4
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.36.2
regex==2023.12.25
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
ruff==0.9.10
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
SpeechRecognition==3.10.0
SQLAlchemy==2.0.25
stack-data==0.6.3
starlette==0.46.1
streamlit==1.43.1
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
text-unidecode==1.3
threadpoolctl==3.5.0
tifffile==2025.2.18
tokenizers==0.21.1
toml==0.10.2
tomlkit==0.13.2
torch==2.6.0
torchvision==0.21.0
tornado==6.3.3
tqdm==4.67.1
traitlets==5.13.0
transformers==4.51.0
typer==0.15.2
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
urllib3==2.1.0
uvicorn==0.34.0
watchdog==6.0.0
wcwidth==0.2.10
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.0.1
widgetsnbextension==4.0.13
wrapt==1.17.2
xxhash==3.5.0
xyzservices==2025.1.0
yarl==1.19.0
zipp==3.18.1
